+++ R/AOI_seq.R
@@ -1,12 +1,12 @@
 #' Sequence analysis of area of interest entries
 #'
-#' Analyses the sequence of entries into defined AOI regions across trials. Can only be used with fixation data with a "fix_n" column denoting fixation events.
-#'
+#' Analyses the sequence of entries into defined AOI regions across trials. Can only be used with fixation data with a "fix_n" column denoting fixation events. 
+#' Assumes that AOIs are non-overlapping and hasn't been tested with overlapping AOIs. Consecutive fixations within an AOI are grouped together as a single entry.
+#' Non-consecutive fixations in the same AOI (i.e., with an intervening fixation in no AOI) are treated as two separate entries. 
 #'
 #' @param data A dataframe with fixation data (from fixation_dispersion). Either single or multi participant data
 #' @param AOIs A dataframe of areas of interest (AOIs), with one row per AOI (x, y, width_radius, height).
 #' @param AOI_names An optional vector of AOI names to replace the default "AOI_1", "AOI_2", etc.
-#' @param participant_ID the variable that determines the participant identifier. If no column present, assumes a single participant
 #' @param progress Display a progress bar
 #' @return a dataframe containing the sequence of entries into AOIs on each trial, entry/exit/duration time into AOI
 #' @export
@@ -14,22 +14,17 @@
 #' @examples
 #' \donttest{
 #' data <- combine_eyes(HCL)
-#' fix_d <- fixation_dispersion(data, participant_ID = "pNum")
+#' fix_d <- fixation_dispersion(data)
 #'
-#' AOI_seq(fix_d, AOIs = HCL_AOIs, participant_ID = "pNum")
+#' AOI_seq(fix_d, AOIs = HCL_AOIs)
 #' }
 #'
 #' @import pbapply
 #' @importFrom stats setNames complete.cases
 
-AOI_seq <- function(data, AOIs, AOI_names = NULL, participant_ID = "participant_ID", progress = TRUE) {
+AOI_seq <- function(data, AOIs, AOI_names = NULL, progress = TRUE) {
 
-  if(is.null(data[["fix_n"]])) stop("column 'fix_n' not detected. Are you sure this is fixation data from eyetools?")
-
-  #first check for multiple/single ppt data
-  test <- .check_ppt_n_in(participant_ID, data)
-  participant_ID <- test[[1]]
-  data <- test[[2]]
+  if(is.null(data[["fix_n"]])) stop("column 'fix_n' not detected. Are you sure you are supplying fixation data from eyetools?")
 
   #internal_AOI_seq carries the per-participant functionality to be wrapped in the lapply for ppt+ setup
   internal_AOI_seq <- function(data, AOIs, AOI_names) {
@@ -39,33 +34,25 @@ AOI_seq <- function(data, AOIs, AOI_names = NULL, participant_ID = "participant_
     data <- do.call("rbind.data.frame", lapply(split(data, data$trial),
                                                     AOI_seq_trial_process,
                                                     AOIs = AOIs,
-                                                    AOI_names,
-                                               participant_ID))
-
-    colnames(data)[1] <- participant_ID #keep same column as entered
-
-
-    #RETURN THE DATA TO THE SAME FORMAT IF SINGLE PPT
-    if (data[[participant_ID]][1] == "NOT A VALID ID") data[[participant_ID]] <- NULL
+                                                    AOI_names))
 
     return(data)
 
   }
 
-  data <- split(data, data[[participant_ID]])
+  data <- split(data, data$pID)
   if(progress) out <- pblapply(data, internal_AOI_seq, AOIs, AOI_names) else out <- lapply(data, internal_AOI_seq, AOIs, AOI_names)
   out <- do.call("rbind.data.frame", out)
   rownames(out) <- NULL
-  out <- .check_ppt_n_out(out)
 
   return(out)
 }
 
 
-AOI_seq_trial_process <- function(trial_data, AOIs, AOI_names, participant_ID) {
+AOI_seq_trial_process <- function(trial_data, AOIs, AOI_names) {
 
   trial_val <- trial_data$trial[[1]]
-  ppt_val <- trial_data[[participant_ID]][[1]]
+  ppt_val <- trial_data$pID[1]
 
   trial_data <- trial_data[complete.cases(trial_data),] # remove any NAs (i.e., in raw data)
 
@@ -91,7 +78,7 @@ AOI_seq_trial_process <- function(trial_data, AOIs, AOI_names, participant_ID) {
   if (sum(aoi_entries)==0) {
 
     # if no data, return a trial result with NAs
-    aoi_trial_out <- data.frame(participant_ID = ppt_val,
+    aoi_trial_out <- data.frame(pID = ppt_val,
                                 trial = trial_val,
                                 AOI = NA,
                                 start = NA,
@@ -106,7 +93,7 @@ AOI_seq_trial_process <- function(trial_data, AOIs, AOI_names, participant_ID) {
 
   # this gives unique values in each row of which AOI had a hit
   aoi_entries <- as.data.frame(as.matrix(aoi_entries)%*%diag(c(1:nrow(AOIs))))
-
+  
   aoi_entries$string <- Reduce(paste0, aoi_entries) # get a string to check for duplicates
 
   aoi_entries$start <- trial_data$start
@@ -124,13 +111,15 @@ AOI_seq_trial_process <- function(trial_data, AOIs, AOI_names, participant_ID) {
   #next section removes duplicate consecutive AOI entries
   aoi_entries <- aoi_entries[!duplicated(with(rle(aoi_entries$string),rep(seq_along(values), lengths))),]
   #remove non AOI region fixations
-  aoi_entries <- aoi_entries[aoi_entries$string != "000",]
+  
+  
+  aoi_entries <- aoi_entries[aoi_entries$string != strrep("0", nrow(AOIs)),]
 
   aoi_entries$AOI <- rowSums(aoi_entries[, -((ncol(aoi_entries) - 3):ncol(aoi_entries))]) # just the AOIs, remove all others
 
 
-  aoi_trial_out <- data.frame(participant_ID = trial_data[[participant_ID]][1],
-                              trial = trial_data$trial[1],
+  aoi_trial_out <- data.frame(pID = ppt_val,
+                              trial = trial_val,
                               AOI = aoi_entries$AOI,
                               start = aoi_entries$start,
                               end = aoi_entries$end,