+++ R/fixation_dispersion.R
@@ -1,26 +1,23 @@
 #' Fixation detection using a dispersion method
 #'
-#' Detects fixations by assessing dispersion of the eye position, using a method that is similar to that proposed by Salvucci & Goldberg (1996).
+#' Detects fixations by assessing dispersion of the eye position, using a method that is similar to that proposed by Salvucci & Goldberg (2000).
 #' Evaluates the maximum dispersion (distance) between x/y coordinates across a window of data. Looks for sufficient periods
 #' in which this maximum dispersion is below the specified dispersion tolerance. NAs are considered breaks
 #' in the data and are not permitted within a valid fixation period.
 #'
-#' It can take either single participant data or multiple participants where there is a variable for unique participant identification.
-#' The function looks for an identifier named `participant_ID` by default and will treat this as multiple-participant data as default,
-#' if not it is handled as single participant data, or the participant_ID needs to be specified
+#' It can take either single participant data or multiple participants, where participants are demarcated by values in the "pID" column.
 #'
 #' @param data A dataframe with raw data (time, x, y, trial) for one participant (the standardised raw data form for eyetools)
 #' @param min_dur Minimum duration (in milliseconds) of period over which fixations are assessed
 #' @param disp_tol Maximum tolerance (in pixels) for the dispersion of values allowed over fixation period
 #' @param NA_tol the proportion of NAs tolerated within any window of samples that is evaluated as a fixation
 #' @param progress Display a progress bar
-#' @param participant_ID the variable that determines the participant identifier. If no column present, assumes a single participant
 #' @return a dataframe containing each detected fixation by trial, with mean x/y position in pixel, start and end times, and duration.
 #' @export
 #' @examples
 #' \donttest{
 #' data <- combine_eyes(HCL)
-#' fixation_dispersion(data, participant_ID = "pNum")
+#' fixation_dispersion(data)
 #' }
 #'
 #' @importFrom utils tail
@@ -28,17 +25,14 @@
 #'
 #' @references Salvucci, D. D., & Goldberg, J. H. (2000). Identifying fixations and saccades in eye-tracking protocols. Proceedings of the Symposium on Eye Tracking Research & Applications - ETRA '00, 71–78.
 
-fixation_dispersion <- function(data, min_dur = 150, disp_tol = 100, NA_tol = .25, progress = TRUE, participant_ID = "participant_ID") {
+fixation_dispersion <- function(data, min_dur = 150, disp_tol = 100, NA_tol = .25, progress = TRUE) {
 
-  #first check for multiple/single ppt data
-  test <- .check_ppt_n_in(participant_ID, data)
-  participant_ID <- test[[1]]
-  data <- test[[2]]
-
+  # check data format
+  .check_data_format(data)
 
   internal_fixation_dispersion <- function(data, min_dur, disp_tol, NA_tol, progress) {
 
-    ppt_label <- data[[participant_ID]][1]
+    ppt_label <- data$pID[1]
 
     data <- split(data,data$trial) # create list from data by trial
     # present a progress bar, unless set to false
@@ -64,7 +58,7 @@ fixation_dispersion <- function(data, min_dur = 150, disp_tol = 100, NA_tol = .2
 
 
     data_fix <- cbind(ppt_label, data_fix)
-    colnames(data_fix)[1] <- participant_ID
+    colnames(data_fix)[1] <- "pID"
 
 
     #row.names(data_fix) <- NULL # remove the row names
@@ -207,12 +201,12 @@ fixation_dispersion <- function(data, min_dur = 150, disp_tol = 100, NA_tol = .2
     }
   }
 
-  data <- split(data, data[[participant_ID]])
+  data <- split(data, data$pID)
   out <- lapply(data, internal_fixation_dispersion, min_dur, disp_tol, NA_tol, progress)
   out <- do.call("rbind.data.frame", out)
   rownames(out) <- NULL
 
-  out <- .check_ppt_n_out(out)
+  # out <- .check_ppt_n_out(out)
 
   return(out)