+++ R/plot_seq.R
@@ -1,13 +1,14 @@
 #' Plot of raw data over time
 #'
-#' A tool for visualising the timecourse of raw data over a single trial. If data from multiple trials are present, then
-#' a single trial will be sampled at random. Alternatively, the trial_number can be specified. Data can be plotted across the whole
-#' trial, or can be split into bins to present distinct plots for each time window.
+#' A tool for visualising the timecourse of raw data over a single trial. Participant and trial values can be selected from the data. 
+#' If values for these parameters are not provided then a single participant and a single trial will be sampled at random. 
+#' Data can be split into bins by time or by the number of bins.
 #'
 #' @param data A dataframe with raw data. If multiple trials are used, then one trial is sampled at random.
-#' @param trial_number can be used to select a particular trial within the data
 #' @param AOIs A dataframe of areas of interest (AOIs), with one row per AOI (x, y, width_radius, height).
-#' @param bg_image The filepath of an image to be added to the plot, for example to show a screenshot of the task.
+#' @param pID_values specify particular values within 'pID' to plot data from certain participants
+#' @param trial_values specify particular values within 'trial' to plot data from certain trials
+#' @param bg_image The filepath of a PNG image to be added to the plot, for example to show a screenshot of the task.
 #' @param res resolution of the display to be shown, as a vector (xmin, xmax, ymin, ymax)
 #' @param flip_y reverse the y axis coordinates (useful if origin is top of the screen)
 #' @param plot_header display the header title text which explains graphical features of the plot.
@@ -22,13 +23,13 @@
 #' data <- combine_eyes(HCL)
 #'
 #' # plot the raw data
-#' plot_seq(data = data[data$pNum == 118,])
+#' plot_seq(data)
 #'
 #' # with AOIs
-#' plot_seq(data = data[data$pNum == 118,], AOIs = HCL_AOIs)
+#' plot_seq(data, AOIs = HCL_AOIs)
 #'
 #' # plot raw data with bins
-#' plot_seq(data = data[data$pNum == 118,], bin_time = 500)
+#' plot_seq(data, bin_time = 500)
 #'
 #' @import ggplot2
 #' @importFrom utils head
@@ -37,8 +38,9 @@
 #'
 
 plot_seq <- function(data = NULL,
-                     trial_number = NULL,
                      AOIs = NULL,
+                     trial_values = NULL,
+                     pID_values = NULL,
                      bg_image = NULL,
                      res = c(0,1920,0,1080),
                      flip_y = FALSE,
@@ -46,26 +48,13 @@ plot_seq <- function(data = NULL,
                      bin_time = NULL,
                      bin_range = NULL) {
 
+  # check pID_values or select random pID
+  data <- .select_pID_values(data, pID_values, allow_random = TRUE)
+  
+  # check trial_values or select random trial
+  data <- .select_trial_values(data, trial_values, allow_random = TRUE)
 
-  # if there is a particular trial number specified
-  if (!is.null(trial_number)){
-    data <- data[data$trial==trial_number,]
-    if (nrow(data)==0){
-      stop("With specified trial number no data found.")
-    }
-
-  } else {
-    # get a random sample from the trial list
-    trial_list <- unique(data$trial)
-
-    if (length(trial_list)>1) {
-      rand_trial <- sample(trial_list,1)
-      message(paste0("Multiple trials detected: randomly sampled - trial:", rand_trial))
-      data <- data[data$trial==rand_trial,]
-    }
-  }
-
-  data$time <- data$time - data$time[1] # start trial timestamps at 0
+  data$time <- data$time - data$time[1] # start timestamps at 0
 
   if (is.null(bin_time)==FALSE){
     data$bin <- ceiling(data$time/bin_time)
@@ -75,9 +64,6 @@ plot_seq <- function(data = NULL,
     }
     bin_end <- data$bin*bin_time
     data$bin_end <- bin_end
-      # data$bin_start <- data$bin_end-bin_time
-      # data$bin_name <- as.factor(paste0(data$bin_start, "-", data$bin_end))
-
   }
 
   final_g <- ggplot(data = data)
@@ -134,10 +120,7 @@ minor_breaks_y <- breaks_y[0:4] + ((res[4]-res[3])/8)
                                              max(final_g$data$time)),
                                   labels = c("start", "end")) +
     theme(legend.position = "bottom",
-          #panel.ontop=TRUE,
-          panel.background = element_rect(fill = NA, colour = NA)) +#,
-         # panel.grid.major = element_line(linewidth = .25),
-          #panel.grid.minor = element_line(linewidth = .125)) +
+          panel.background = element_rect(fill = NA, colour = NA)) +
     labs(y = "Vertical coordinate (pixels)",
          x = "Horizontal coordinate (pixels)")
 
@@ -146,7 +129,7 @@ minor_breaks_y <- breaks_y[0:4] + ((res[4]-res[3])/8)
     final_g <-
       final_g +
       labs(title = "eyetools::plot_seq()",
-           subtitle = "Raw data shown as dots; \nAOIs shown as blue regions")
+           subtitle = "Raw data shown as X; \nAOIs shown as blue regions")
   }
 
   # setting axes limits and reversing y
@@ -167,7 +150,7 @@ minor_breaks_y <- breaks_y[0:4] + ((res[4]-res[3])/8)
 
 # function to add raw data
 add_raw_time_seq <- function(dataIn, ggplot_in){
-
+  
   x <- dataIn$x
   y <- dataIn$y
 
@@ -175,7 +158,7 @@ add_raw_time_seq <- function(dataIn, ggplot_in){
     ggplot_in +
     geom_point(data = dataIn,
                aes(x = x, y = y, colour = time),
-               shape = 4,
+               shape = 16,
                size = 3,
                alpha = .5,
                na.rm = TRUE)